+++ libraries/float16/float16.cpp
@@ -1,7 +1,7 @@
 //
 //    FILE: float16.cpp
 //  AUTHOR: Rob Tillaart
-// VERSION: 0.1.8
+// VERSION: 0.2.0
 // PURPOSE: library for Float16s for Arduino
 //     URL: http://en.wikipedia.org/wiki/Half-precision_floating-point_format
 
@@ -191,23 +191,18 @@ float float16::f16tof32(uint16_t _value) const
     else return NAN;
   }
 
-  //  SUBNORMAL/NORMAL
-  if (exp == 0)  f = 0;
-  else           f = 1;
-
-  //  PROCESS MANTISSE
-  for (int i = 9; i >= 0; i--)
+  //  NORMAL
+  if (exp > 0)
   {
-    f *= 2;
-    if (man & (1 << i)) f = f + 1;
+    f = pow(2.0, exp - 15) * (1 + man * 0.0009765625);
+    return sgn ? -f : f;
   }
-  f = f * pow(2.0, exp - 25);
-  if (exp == 0)
-  {
-    f = f * pow(2.0, -13);    // 5.96046447754e-8;
-  }
+  //  SUBNORMAL
+  //  exp == 0;
+  f = pow(2.0, -24) * man;
   return sgn ? -f : f;
 }
+
 
 uint16_t float16::f32tof16(float f) const
 {
@@ -217,50 +212,76 @@ uint16_t float16::f32tof16(float f) const
   int16_t  exp = (t & 0x7F800000) >> 23;
   bool     sgn = (t & 0x80000000);
 
+// Serial.print("BEFOR:\t ");
+// Serial.print(sgn, HEX);
+// Serial.print(" ");
+// Serial.print(man, HEX);
+// Serial.print(" ");
+// Serial.println(exp, HEX);
+
   //  handle 0
   if ((t & 0x7FFFFFFF) == 0)
   {
     return sgn ? 0x8000 : 0x0000;
   }
+
   //  denormalized float32 does not fit in float16
   if (exp == 0x00)
   {
     return sgn ? 0x8000 : 0x0000;
   }
-  //  handle infinity & NAN
+
+  //  handle INF and NAN == infinity and not a number
   if (exp == 0x00FF)
   {
     if (man) return 0xFE00;         //  NAN
     return sgn ? 0xFC00 : 0x7C00;   //  -INF : INF
   }
 
-  //  normal numbers
+  //  rescale exponent
   exp = exp - 127 + 15;
-  //  overflow does not fit => INF
+
+  //  overflow does not fit => INF (infinity)
   if (exp > 30)
   {
     return sgn ? 0xFC00 : 0x7C00;   //  -INF : INF
   }
-  //  subnormal numbers
-  if (exp < -38)
+
+  //  subnormal numbers out of range => 0.
+  if (exp < -9)
   {
     return sgn ? 0x8000 : 0x0000;   //  -0 or 0  ?   just 0 ?
   }
-  if (exp <= 0) //  subnormal
+
+  //  subnormal numbers
+  if (exp <= 0)
   {
-    man >>= (exp + 14);
-    //  rounding
-    man++;
-    man >>= 1;
+    exp = 0;
+    man = abs(f) * 16777216;  //  pow(2.0, 24);
     if (sgn) return 0x8000 | man;
     return man;
   }
 
-  //  normal
-  //  TODO rounding
-  exp <<= 10;
+
+  //  normal numbers
+  //  rounding
   man++;
   man >>= 1;
+  //  correction mantissa overflow issue #10
+  if (man == 0x0400)
+  {
+    exp++;
+    man = 0;
+  }
+  exp <<= 10;
+
+// Serial.print("AFTER:\t ");
+// Serial.print(sgn, HEX);
+// Serial.print(" ");
+// Serial.print(man, HEX);
+// Serial.print(" ");
+// Serial.println(exp, HEX);
+
   if (sgn) return 0x8000 | exp | man;
   return exp | man;
 }