+++ libraries/float16/float16.h
@@ -2,7 +2,7 @@
 //
 //    FILE: float16.h
 //  AUTHOR: Rob Tillaart
-// VERSION: 0.1.8
+// VERSION: 0.2.0
 // PURPOSE: Arduino library to implement float16 data type.
 //          half-precision floating point format,
 //          used for efficient storage and transport.
@@ -11,7 +11,7 @@
 
 #include "Arduino.h"
 
-#define FLOAT16_LIB_VERSION                 (F("0.1.8"))
+#define FLOAT16_LIB_VERSION                 (F("0.2.0"))
 
 
 class float16: public Printable
@@ -34,22 +34,22 @@ public:
   uint8_t  getDecimals()          { return _decimals; };
 
   //  equalities
-  bool operator == (const float16& f);
-  bool operator != (const float16& f);
+  bool     operator == (const float16& f);
+  bool     operator != (const float16& f);
 
-  bool operator >  (const float16& f);
-  bool operator >= (const float16& f);
-  bool operator <  (const float16& f);
-  bool operator <= (const float16& f);
+  bool     operator >  (const float16& f);
+  bool     operator >= (const float16& f);
+  bool     operator <  (const float16& f);
+  bool     operator <= (const float16& f);
 
   //  negation
-  float16 operator - ();
+  float16  operator - ();
 
   //  basic math
-  float16 operator + (const float16& f);
-  float16 operator - (const float16& f);
-  float16 operator * (const float16& f);
-  float16 operator / (const float16& f);
+  float16  operator + (const float16& f);
+  float16  operator - (const float16& f);
+  float16  operator * (const float16& f);
+  float16  operator / (const float16& f);
 
   float16& operator += (const float16& f);
   float16& operator -= (const float16& f);
@@ -57,14 +57,14 @@ public:
   float16& operator /= (const float16& f);
 
   //  math helper functions
-  int   sign();       //  1 = positive   0 = zero   -1 = negative.
-  bool  isZero();
-  bool isNaN();
-  bool isInf();
+  int      sign();       //  1 = positive   0 = zero   -1 = negative.
+  bool     isZero();
+  bool     isNaN();
+  bool     isInf();
 
 
   //  CORE CONVERSION
-  //  should be private but for testing...
+  //  should be private, needed for testing.
   float    f16tof32(uint16_t) const;
   uint16_t f32tof16(float) const;